fix issue 864
